server/requirements_msix.txt
# MSIX Packaging Requirements
# These versions are compatible with Windows Store requirements

# Core GUI framework
customtkinter>=5.2.0
tkinter

# Web framework
Flask>=3.0.0
Werkzeug>=3.0.0
Jinja2>=3.0.0

# AI/ML libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
safetensors>=0.3.0

# Image and audio processing
Pillow>=9.0.0
opencv-python>=4.8.0
librosa>=0.10.0
soundfile>=0.12.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# PDF processing
pdf2image>=1.16.0
pytesseract>=0.3.10
pypdf==6.0.0

# Utilities
requests>=2.31.0
PyYAML>=6.0
python-dateutil>=2.8.0

# MSIX specific
pyinstaller>=5.0.0
